Change IntensityTable x,y,z SpotAttributes to floats
Spot detection methods return spot coordinates with
sub-pixel resolution. Previously this sub-pixel information
was thrown away due to int conversion (see: #898).

The correct behavior should be to retain sub-pixel
information and modify downstream functions to
properly handle float values for x,y,z pixel
coordinates.

Relates to: #1310

+++ starfish/core/spots/_detect_spots/blob.py
@@ -114,8 +114,8 @@ class BlobDetector(DetectSpotsAlgorithmBase):
         converted_radius = np.round(fitted_blobs[Features.SPOT_RADIUS] * np.sqrt(3))
         fitted_blobs[Features.SPOT_RADIUS] = converted_radius
 
-        # convert the array to int so it can be used to index
-        rounded_blobs = SpotAttributes(fitted_blobs.astype(int))
+        # Fitted blob attributes need to be kept as floats!
+        rounded_blobs = SpotAttributes(fitted_blobs.astype(float))
 
         rounded_blobs.data['intensity'] = measure_spot_intensity(
             data_image, rounded_blobs, self.measurement_function)
