+++ starfish/core/spots/_detect_spots/trackpy_local_max_peak_finder.py
@@ -137,10 +137,8 @@ class TrackpyLocalMaxPeakFinder(DetectSpotsAlgorithmBase):
             attributes.columns = new_colnames
 
         attributes['spot_id'] = np.arange(attributes.shape[0])
-        # convert these to int so it can be used to index
-        attributes.x = attributes.x.astype(int)
-        attributes.y = attributes.y.astype(int)
-        attributes.z = attributes.z.astype(int)
+
+        # Spot attributes need to be kept as floats!
         return SpotAttributes(attributes)
 
     def run(
